+ adapter + simple factory comment + flyweight
